mysite/requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-postman==3.6.2
django-rosetta==0.9.4
gunicorn==20.0.4
idna==2.9
multidict==4.7.5
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.2.0
polib==1.1.0
psycopg2==2.8.5
PyMySQL==0.9.3
python-engineio==3.12.0
python-gettext==4.0
python-socketio==4.5.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
stripe==2.43.0
urllib3==1.25.8
whitenoise==5.1.0
yarl==1.4.2
django-toolbelt==0.0.1
django-model-utils==4.0.0
pinax-messages==3.0.0
pinax-templates==2.0.6
wiki==0.6
django-sekizai==2.0.0
django-mptt==0.11.0
django-nyt==1.1.5





